workaround for on() req in game.tsx - currently broken
